// apps/api/src/modules/auth/schemas.ts
import { z } from 'zod'

export const RegisterSchema = z.object({
  email: z.string().email(),
  password: z.string().min(8)
})

export const LoginSchema = z.object({
  email: z.string().email(),
  password: z.string().min(8)
})

export const RefreshSchema = z.object({
  refreshToken: z.string().min(10)
})